Resolved rollback issue
