+++ admin-core/assets/src/settings-app/pages/BlocksPage.js
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react';
 import apiFetch from '@wordpress/api-fetch';
 import { useStateValue } from '@Utils/StateProvider';
 import './BlocksPage.scss';
-import { __ } from '@wordpress/i18n';
 import BlocksSettings from '@SettingsApp/components/blocks-page/BlocksSettings';
 import DisplayCondition from '@SettingsApp/components/blocks-page/DisplayCondition';
 import SettingPageSkeleton from '@Admin/settings-app/components/blocks-page/BlocksPageSkeleton';
